requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
opencv-python>=4.5.3
Pillow>=8.3.1
matplotlib>=3.4.2
tqdm>=4.61.2
scikit-image>=0.18.1